+++ src/scl/utility.h
@@ -5,6 +5,7 @@
  * @brief Includes all ScL Utility headers.
  */
 
+#include <scl/utility/any.h>
 #include <scl/utility/attribute.h>
 #include <scl/utility/concepts.h>
 #include <scl/utility/flags.h>